code-file
The traffic shaper will queue the incoming packets from A and then schedule them at a constant bit rate towards Station-B. Thus, the traffic between Station A and the Traffic Shaper is a variable bitrate, while between the Traffic Shaper and Station-B is constant bitrate. The queue inside the Traffic Shaper would be a FIFO queue with a fixed size and constant service time.